Improve the performance of the solution to day 12 for 2021.

## Before

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1348 (21H1/May2021Update)
Intel Core i9-9980HK CPU 2.40GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.100
  [Host]     : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
  DefaultJob : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT


```
| Method |    input |     Mean |   Error |  StdDev |     Gen 0 | Allocated |
|------- |--------- |---------:|--------:|--------:|----------:|----------:|
|  Solve | Y2021-12 | 173.3 ms | 3.36 ms | 3.45 ms | 8333.3333 |     68 MB |

## After

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1348 (21H1/May2021Update)
Intel Core i9-9980HK CPU 2.40GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.100
  [Host]     : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
  DefaultJob : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT


```
| Method |    input |     Mean |   Error |  StdDev |     Gen 0 | Allocated |
|------- |--------- |---------:|--------:|--------:|----------:|----------:|
|  Solve | Y2021-12 | 107.7 ms | 2.13 ms | 1.99 ms | 2400.0000 |     19 MB |